fix : 회원 정보 관련 기능 추가

+++ src/main/java/efub/toy2/papers/domain/comment/controller/CommentController.java
@@ -4,7 +4,10 @@ import efub.toy2.papers.domain.comment.dto.CommentRequestDto;
 import efub.toy2.papers.domain.comment.dto.CommentResponseDto;
 import efub.toy2.papers.domain.comment.service.CommentService;
 import efub.toy2.papers.domain.member.domain.Member;
+import efub.toy2.papers.domain.member.service.MemberService;
 import efub.toy2.papers.global.config.AuthUser;
+import efub.toy2.papers.global.exception.CustomException;
+import efub.toy2.papers.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -17,17 +20,20 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CommentController {
     private final CommentService commentService;
+    private final MemberService memberService;
 
 
     /* 댓글 생성 */
     @PostMapping("/comments")
     public CommentResponseDto createComment(@AuthUser Member member, @RequestBody CommentRequestDto requestDto){
+        if(!memberService.isAdminMember(member)) throw new CustomException(ErrorCode.NON_LOGIN);
         return commentService.createComment(member,requestDto);
     }
 
     /* 댓글 삭제 */
     @DeleteMapping("/comments/{commentId}")
     public String deleteComment(@AuthUser Member member ,@PathVariable Long commentId){
+        if(!memberService.isAdminMember(member)) throw new CustomException(ErrorCode.NON_LOGIN);
         return commentService.deleteComment(member,commentId);
     }
